Split component files in extension generator
